+++ CHANGES.md
@@ -1,3 +1,4 @@
 ### App
+- Added Button in Automation View to skip to Turnaround Phase (only available in Preparation Phase)
 - Remove 'Stand' from captured Gate
 - Added some extra Log Messages